QE-1 Решить квадратное уравнение ax^2+bx+c=-7. На вход в программу подаются значения a,b,c. На выходе хотим получить решение уравнения ( в процессе )

+++ Krujka/src/main/java/com.company/Krujka.java
@@ -3,10 +3,10 @@ package com.company;
 class Krujka {
     public static void main(String[] args) {
 
-        int a = 1;
-        int b = 2;
+        double a = 1.0;
+        double b = 2.0;
         double c = 2.75;
-        int d = (int) (b * 2 - 4 * a * c);
+        double d = b * 2 - 4 * a * c;
         System.out.print(d);
     }
 }
